+++ app/Http/Controllers/Admin/LeagueController.php
@@ -6,6 +6,7 @@ use DB;
 use Exception;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\Request;
+use App\Models\League;
 use App\Http\Requests\LeagueRequest;
 use App\Exception\RepositoryException;
 use App\Http\Controllers\Controller;
@@ -27,6 +28,7 @@ class LeagueController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('access', League::class);
         // q: query parameter
         $keyword = $request->q;
 
@@ -34,7 +36,7 @@ class LeagueController extends Controller
             $leagues = $this->leagueRepository->search($keyword, config('repository.pagination.limit'));
             $leagues->appends($request->only('q'));
         } else {
-            $leagues = $this->leagueRepository->leagues(config('repository.pagination.limit'));
+            $leagues = $this->leagueRepository->leagues(config('repository.pagination.limit'), [['id', 'desc']]);
         }
 
         return view('admin.league.index', compact('leagues'));
@@ -48,6 +50,7 @@ class LeagueController extends Controller
      */
     public function store(LeagueRequest $request)
     {
+        $this->authorize('access', League::class);
         try {
             $this->leagueRepository->create($request->all());
 
@@ -75,9 +78,10 @@ class LeagueController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('access', League::class);
         try {
             $league = $this->leagueRepository->find($id); // throw RepositoryException when can not found
-            $leagues = $this->leagueRepository->leagues(config('repository.pagination.limit'));
+            $leagues = $this->leagueRepository->leagues(config('repository.pagination.limit'), [['id', 'desc']]);
 
             return view('admin.league.index', compact('league', 'leagues'));
         } catch (RepositoryException $e) {
@@ -100,6 +104,7 @@ class LeagueController extends Controller
      */
     public function update(LeagueRequest $request, $id)
     {
+        $this->authorize('access', League::class);
         try {
             $league = $this->leagueRepository->find($id); // throw RepositoryException when can not found
             $this->leagueRepository->update($request->all(), $league);
@@ -129,6 +134,7 @@ class LeagueController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('access', League::class);
         try {
             $this->leagueRepository->find($id); // throw RepositoryException when can not found
             $this->leagueRepository->delete($id);
@@ -152,6 +158,10 @@ class LeagueController extends Controller
             ];
         }
 
-        return redirect()->route('leagues.index')->with('notification', $notification);
+        if (str_contains(url()->previous(), route('leagues.edit', ['id' => $id]))) {
+            return redirect()->route('leagues.index')->with('notification', $notification);
+        }
+
+        return redirect()->back()->with('notification', $notification);
     }
 }
